Change float into double

+++ src/main/kotlin/io/github/yuuuu51/dijkstralib/element/Edge.kt
@@ -2,7 +2,7 @@ package io.github.yuuuu51.dijkstralib.element
 
 class Edge(
     val id: String,
-    val length: Float
+    val length: Double
 ) : Element {
 
     val nextNodes = mutableListOf<Node>()
